chore(tests): split up unissued facilities e2e tests
