+++ e2e-tests/gef/cypress/e2e/facility/pre-submission/about-facility.spec.js
@@ -1,10 +1,10 @@
-import relative from '../relativeURL';
-import { backLink, continueButton, errorSummary, headingCaption, mainHeading, form, saveAndReturnButton } from '../partials';
-import aboutFacility from '../pages/about-facility';
-import bankReviewDate from '../pages/bank-review-date';
-import facilityEndDate from '../pages/facility-end-date';
-import { BANK1_MAKER1 } from '../../../../e2e-fixtures/portal-users.fixture';
-import dateConstants from '../../../../e2e-fixtures/dateConstants';
+import relative from '../../relativeURL';
+import { backLink, continueButton, errorSummary, headingCaption, mainHeading, form, saveAndReturnButton } from '../../partials';
+import aboutFacility from '../../pages/about-facility';
+import bankReviewDate from '../../pages/bank-review-date';
+import facilityEndDate from '../../pages/facility-end-date';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
+import dateConstants from '../../../../../e2e-fixtures/dateConstants';
 
 const applications = [];
 let token;
