+++ e2e-tests/gef/cypress/e2e/facility/pre-submission/facility-confirm-deletion.spec.js
@@ -1,8 +1,8 @@
-import relative from '../relativeURL';
-import { mainHeading } from '../partials';
-import facilityConfirmDeletion from '../pages/facility-confirm-deletion';
-import applicationDetails from '../pages/application-details';
-import { BANK1_MAKER1 } from '../../../../e2e-fixtures/portal-users.fixture';
+import relative from '../../relativeURL';
+import { mainHeading } from '../../partials';
+import facilityConfirmDeletion from '../../pages/facility-confirm-deletion';
+import applicationDetails from '../../pages/application-details';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
 
 const applications = [];
 let token;
