+++ e2e-tests/gef/cypress/e2e/facility/pre-submission/facility-value.spec.js
@@ -1,7 +1,7 @@
-import relative from '../relativeURL';
-import { backLink, errorSummary, headingCaption, mainHeading, continueButton } from '../partials';
-import facilityValuePage from '../pages/facility-value';
-import { BANK1_MAKER1 } from '../../../../e2e-fixtures/portal-users.fixture';
+import relative from '../../relativeURL';
+import { backLink, errorSummary, headingCaption, mainHeading, continueButton } from '../../partials';
+import facilityValuePage from '../../pages/facility-value';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
 
 const applications = [];
 let token;
